test: move the test case to test_tenant_deletion.py
